show list

+++ stopRPi.pl
@@ -5,7 +5,8 @@ use warnings;
 my @pidList;
 my @ProcList = ('SystemMonitor.pl','MapNetwork.pl','WatchDog.pl','ConfigServer.pl','ShutdownNUC.pl','i2c_apc_v6.pl',
 				'RemoteLog.pl', 'RunNVR.pl', 'RemoteData.pl','WatchSSD.pl','i2c_apc_v6.pl','Watcher.pl','WatchRPi.pl');
-`ps -elf |grep perl`;
+my $ret = `ps -elf |grep perl`;
+print $ret;
 foreach my $curProc (@ProcList) {
 	my $return = `ps -C $curProc -o pid`;
  	my @lines = split("\n",$return);
@@ -17,7 +18,11 @@ foreach my $curProc (@ProcList) {
  	}
 	push(@pidList,$pid) unless $pid == -1;
 }
-my $cmd = "kill " . join(" ",@pidList);
-`$cmd`;
-`ps -elf |grep perl`;
+if (scalar @pidList > 0) {
+	my $cmd = "kill " . join(" ",@pidList);
+	`$cmd`;	
+}
+$ret = `ps -elf |grep perl`;
+print $ret;
+
 1;
